+++ src/App.ts
@@ -1,14 +1,16 @@
 import * as Document from "./Document";
 import * as NS from "./NS";
 import * as ObjectSchema from "./ObjectSchema";
+import Pointer from "./Pointer";
 import * as Utils from "./Utils";
-import {IllegalArgumentError} from "./Errors";
+import IllegalArgumentError from "./Errors/IllegalArgumentError";
 
 import Context from "./App/Context";
 
 export interface Class extends Document.Class {
 	name:string;
 	description?:string;
+	allowsOrigins:(Pointer | string)[];
 }
 
 export const RDF_CLASS:string = NS.CS.Class.Application;